feat: simplifies Tooltip Component, creates Abbr component, updates Date and Range Pickers
