+++ src/components/Tooltip/index.ts
@@ -1,2 +1,1 @@
 export * from "./Tooltip";
-export * from "./WithTooltip";
