+++ src/components/index.ts
@@ -1,3 +1,4 @@
+export * from "./Abbr";
 export * from "./Avatar";
 export * from "./Badge";
 export * from "./Breadcrumbs";
